fsbs-platform/sidecar/requirements.txt
# OTLP protocol — to receive and forward spans
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-exporter-otlp-proto-grpc==1.28.0
opentelemetry-proto==1.28.0

# gRPC server — sidecar listens for OTLP spans
grpcio==1.68.0
grpcio-tools==1.68.0
protobuf==5.28.0

# Math — small matrix operations for LinUCB
numpy==2.1.0

# Testing
pytest==8.3.0